Revert "Admin auth"
