--- fn/admin/src/actions/Auth-actions.js
@@ -1,6 +1,0 @@
-const setAdmin = data => ({
-    type: 'SET_ADMIN',
-    payload: data,
-});
-
-export { setAdmin };
