--- fn/admin/src/components/login-input/Login-input-styles.js
@@ -1,7 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-export const useStyles = makeStyles(() => ({
-    input: {
-        maxWidth: '300px'
-    }
-}));
