--- fn/admin/src/components/login-input/index.js
@@ -1,3 +1,0 @@
-import LoginInput from './Login-input';
-
-export default LoginInput;
