--- fn/admin/src/components/login-page/index.js
@@ -1,3 +1,0 @@
-import LoginPage from './Login-page';
-
-export default LoginPage;
