+++ fn/admin/src/config/index.js
@@ -17,7 +17,6 @@ const routes = {
     pathToAddBrand: '/brandadd',
     pathToUsers: '/users',
     pathToUserDetails: '/user/:id',
-    pathToLogin: '/login',
 };
 
 export const config = {
@@ -31,7 +30,7 @@ export const config = {
             ['Users', routes.pathToUsers, PeopleAltIcon],
         ],
         routes,
-        serverUrl: 'http://localhost:5000/',
+        serverUrl: 'https://lv487node-backend.herokuapp.com/',
         drawerWidth: 220,
         snackBarDuration: 4000,
         rowsPerPageOptions: [10, 25, 50, 100],
