+++ fn/admin/src/container/index.js
@@ -6,7 +6,6 @@ import BrandsPageContainer from './brand-page-container';
 import CategoriesPageContainer from './categories-page-container';
 import CategoryDetailsPageContainer from './category-details-page-container';
 import ProductDetailsPage from './product-details-container';
-import LoginPageContainer from './login-page-container';
 
 export {
     ProductsPageContainer,
@@ -17,5 +16,4 @@ export {
     CategoriesPageContainer,
     CategoryDetailsPageContainer,
     ProductDetailsPage,
-    LoginPageContainer,
 };
