--- fn/admin/src/container/login-page-container/index.js
@@ -1,3 +1,0 @@
-import LoginPageContainer from './Login-page-container';
-
-export default LoginPageContainer;
