+++ fn/admin/src/services/Admin-service.js
@@ -19,7 +19,7 @@ export default class AdminService {
             const response = await axios.post(`${serverUrl}${url}`, dataToSend);
             return response.data;
         } catch (error) {
-            throw error;
+            console.error(error);
         }
     };
