+++ fn/admin/src/services/Users-service.js
@@ -20,15 +20,6 @@ class UsersService extends AdminService {
         const res = await this.putData(`users/role/${user._id}`, { user });
         return res;
     };
-
-    loginAdmin = async user => {
-        try {
-            const res = await this.postData(`auth/admin/login`, user);
-            return res;
-        } catch (err) {
-            throw err;
-        }
-    }
 }
 
 const usersService = new UsersService();
