+++ fn/middleware/auth.js
@@ -1,8 +1,6 @@
 const jwt = require('jsonwebtoken');
-const User = require('../models/User');
-require('dotenv').config();
 
-exports.auth = async (req, res, next) => {
+module.exports = function (req, res, next) {
     const token = req.header('x-auth-token');
 
     if (!token) {
@@ -11,21 +9,10 @@ exports.auth = async (req, res, next) => {
 
     try {
         const decoded = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
-        const user = await User.findOne({ email: decoded.name })
-        req.user = user;
+
+        req.user = decoded.user;
         next();
     } catch (error) {
         res.status(401).json({ msg: 'Token is not valid' });
     }
 };
-
-exports.authorize = (...roles) => {
-    return (req, res, next) => {
-        const { role } = req.user
-        if (!roles.includes(role)) {
-            return res.status(403)
-                .send({ msg: `Role ${req.user.role} is not authorized to access this route` });
-        }
-        next();
-    }
-}
