+++ fn/server.js
@@ -6,7 +6,10 @@ const connectDB = require('./config/db');
 const path = require('path');
 const rfs = require('rotating-file-stream');
 
-const errorHandler = require('./middleware/errorHandler');
+const accessLogStream = rfs.createStream('access.log', {
+    interval: '3d',
+    path: path.join(__dirname, 'logs')
+});
 
 const auth = require('./routes/users/auth');
 const users = require('./routes/users/users');
@@ -17,13 +20,6 @@ const brands = require('./routes/products/brands');
 const colors = require('./routes/products/colors');
 const generator = require('./routes/products/generator');
 const orders = require('./routes/purchase/order');
-
-require('dotenv').config();
-
-const accessLogStream = rfs.createStream('access.log', {
-    interval: '3d',
-    path: path.join(__dirname, 'logs')
-});
 
 const app = express();
 app.use(cors());
@@ -45,7 +41,6 @@ app.use('/colors', colors);
 app.use('/generator', generator);
 app.use('/orders', orders);
 
-app.use(errorHandler);
 
 const PORT = process.env.PORT || 5000;
