--- fn/utils/errorResponse.js
@@ -1,8 +1,0 @@
-class ErrorResponse extends Error {
-    constructor(message, statusCode) {
-        super(message);
-        this.statusCode = statusCode;
-    }
-}
-
-module.exports = ErrorResponse;
